Give priority to spec mounts that conflict with sysbox supplementary mounts.
If a sys container spec has a mount entry whose destination conflicts
with a supplementary mount (e.g., /var/lib/docker), give priority
to the supplementary mount.

For example, this allows users to create sys containers with Docker
and mount "/var/lib/docker" with a persistent volume on the host.

This change also fixes a bug that was affecting several functions
in spec.go, and creates unit tests for each of those functions.