Migrate from Nestybox shiftfs to Ubuntu Shiftfs.
This change migrates sysbox-runc from using the Nestybox
shiftfs module to using the Ubuntu shiftfs module.

This migration was done for several reasons:

1) Ubuntu's shiftfs is more complete than the Nestybox shiftfs.

2) We found a way to use the Ubuntu shiftfs mark & mount approach
  with sysbox-runc (see below for details).

3) By using Ubuntu's shiftfs, we no longer have to package the Nestybox
  shiftfs with Sysbox. This simplifies the Sysbox installer and
  enables Sysbox to be used by customers which have a
  "no-external-kernel-module" policy.

4) The Ubuntu shiftfs works on the same distros as the Nestybox
  shiftfs, so there is no disadvantage in using the Ubuntu shiftfs
  module.

A few more details on (2) above:

The Ubuntu shiftfs module uses a two-step mark & mount approach, just
like James Bottomley's original shiftfs did.

We previously we had problems integrating this mark & mount approach
with sysbox-runc, because the mounting step must be done from within
the user-namespace of the system container at which point the system
container may no longer have permission to access the mount points on
the host. This problem was showing up when using Docker + Sysbox,
because Docker places the container rootfs under "/var/lib/docker"
which is only accessible to the host root.

This problem was one of the main reasons we developed the Nestybox
shiftfs module (which changed the two step mark-mount sequence
into a single step mount sequence done from outside the container's
user-ns).

However, we have found a way to make the two-step mark-mount sequence
used by the Ubuntu shiftfs module with sysbox-runc. The trick was to
delay the user-id and group-id mappings for the system container's
user-namespace until after the shiftfs mounts take place. This
required low-level changes in the sysbox-runc libcontainer nsenter
package, as well as in the parts of the libcontainer package that do
the setup of the system conatiner's rootfs.