Fixed bug in cfgSystemd() and added unit tests.
Fixed a bug in function cfgSystemd(), in the code that ensures
systemd env var are present (prior to this change said code
was checking for a full match on a string "container=private-users",
when it should only check for a match on the variable name "container".

I also added a unit test to check that cfgSystemd() works.