Mount kernel headers inside the sys container.
In order to support more system-level apps (e.g,. kubeadm),
it's necessary to mount the kernel header sources into
the sys container.

This change does this.

The change is a bit complex because the host directory that has the
kernel sources may contain symlinks to other directories, and those
need to be mounted inside the sys container too.

Bonus: replaced deprecated filepath.HasPrefix() with strings.HasPrefix().