Cleanup sys container mount configuration.
This change cleans up the sys container mount config
code. It ensures sysbox-runc removes spec mounts
that conflict with its "must-have" mounts, ensures
that mounts are added to the container's spec in the correct order,
and ensures that procfs, sysfs, and /dev and always mounted
in a sys container.

For sysfs, it ensures that it's mounted read-write on /sys.
While mounting sysfs read-write may seem risky, it's fine
because the sys container processes won't be able to write
to non-namespaced files under /sys by virtue of the protection
afforded by the Linux user namespace.

Mouting sysfs read-write is necessary in order to run
privileged containers inside a sys container.