Revert "Merge pull request #2280 from kolyshkin/errors-unwrap"
Using errors.Unwrap() is not the best thing to do, since it returns
nil in case of an error which was not wrapped. More to say,
errors package provides more elegant ways to check for underlying
errors, such as errors.As() and errors.Is().

This reverts commit 7b67216be1fd34fc177179795b0716ffa50b06f3, reversing
changes made to 1aa5a8ef64331f29fa4012656c9f24382c442995.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>