Request the sysbox-mgr to mount a host dir to back a sys container's /var/lib/containerd.
In sys containers that use containerd to pull inner container images,
these images are unpacked into a sub-dir of /var/lib/containerd.

Turns out that the unpacking operation fails if shiftfs is mounted on
/var/lib/containerd (see Sysbox issue #622 for details). To
work-around this, this commit causes sysbox-runc to request the
sysbox-mgr to back the sys container's /var/lib/containerd directory
with a directory on the host. This way, the mount of shiftfs is
avoided.