Setup a host volume to back the sys container's containerd workspace.
In order for containerd inside a sys container to operate properly,
the directory that containerd uses to store overlayfs image snapshots
must be backed by a host directory (i.e., which is not on overlayfs and
does not use shiftfs). This change causes sysbox-runc request the
sysbox-mgr to setup such a directory.