Re-work sysbox-runc nsenter sequence to fix sysbox issue #520.
This change reworks the sysbox-runc nsenter package to prepare
rootfs mounts when a new container is created.

Long story short:

In the OCI runc, when a new container is created, the creation of the
container's namespaces and setup of user-ns uid-mappings is done in
nsenter. The ensuing go-runtime then simply sets up the container
rootfs prior to doing exec into the container's init process.

We modified this some time back to not do the uid-mappings in nsenter
and instead do them in the go-runtime, after preparing the
rootfs. This was done to get around the problem that once the user-ns
uid-mappings are set, we would loose accesss to perform some shiftfs
mounts (e.g., mounting shiftfs bind sources outside of the container's
rootfs).

But this created another problem: files in /proc/<pid>/net showed up
with nobody:nouser inside the container, causing programs like
iptable-save to fail inside the container. The reason was that since
we had delayed the user-ns uid mappings, the mappings were not set
when we entered the network namespace in nsenter, and this confused
the network namespace.

This change fixes this last problem. It does so by moving the rootfs
prep mounts from standard_init_linux.go into nsenter itself. This way,
they can be done prior to setting the user-ns uid-mappings (such that
we don't loose permission to setup shiftfs), and in addition the
uid-mappings can be done before entering the network namespace
(solving the /proc/pid/net ownership problem).