Modify IP address of Docker DNS inside the sys container.
This change ensures that all processes inside the sys container,
including inner containers, access the DNS server programmed
by the host Docker for the sys container.

The change does this by aliasing the IP address of the Docker DNS inside the sys
container, from 127.0.0.11:53 to <container default gateway>:53.

This ensures that when inner Docker containers are launched, the resolv.conf
file in inner containers points to the Docker DNS inside the parent system
container, instead of another DNS (e.g., Googles DNS at 8.8.8.8).

This also helps when running K8s inside system containers: it causes the kubelet
to configure the resolv.conf in kube-system pods (e.g., coredns pod) such that
it points to the Docker DNS in the sys container, as opposed to 127.0.0.11:53
(which would be wrong, and would create resolution-loop in the coredns pods).

Refer to Sysbox issue #679 for full details.