Skip unit tests for detached pass-through TTY.
These tests fails when run under bats, which implicitly redirects stdout of the
test to a log file.

For example, when the test executes "sysbox-runc create test_busybox",
this is really "sysbox-runc create test_busybox >> bats_log_file".

The stdout redirect causes sysbox-runc to change the ownership of the
bats_log_file to the uid(gid) of the sys container's root process. Later when
bats tries to write more logs to the file, it fails due to "permission denied".

It's not clear to me if there's a better fix. Sysbox issue #707 is tracking this.