Fix for issue preventing sysbox-runc test from completing
Affected test-case is verifying the proper operation of cgroups within a system
container. The container being generated is demanding more of memory (just ~4MB)
than what was initially configured as memory limit in container's
mem-cgroup. Thereby, kernel ends up oom-killing processes within the container.

The solution is to make a minimal change in the testcase to increase (5MB) the
cgroup memory-limit of the tested container.

Fixes Sysbox issue #30.

Signed-off-by: Rodny Molina <rmolina@nestybox.com>