Makefile: fix vendor and verify-dependencies
Commit 7c2f2bf88 added these targets. Alas, the `go mod tidy` never
worked, as it was written as part of `export` statement:

	export GO111MODULE=on \
        $(GO) mod tidy && \
	...

which is the same as

	export GO111MODULE=on $(GO) mod tidy && ...

which exports a bunch of variables, such as `go`, `mod`, and `tidy`,
but does not run it.

The fix would be to add a semicolon after the `export` statement,
but since GO111MODULE is not really needed here (maybe some older
golang versions needed it?), let's just drop it.

With this dropped, && does not make any sense, so drop it, too.

NOTE that if someone tries

	GO111MODULE=off make vendor

it will fail, but I guess it is expected.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>