Remove seccomp syscall argument restrictions for sys containers.
The OCI spec allows container managers to use seccomp to restrict syscalls
inside containers. Furthermore, it's possible to restrict syscalls based on
arguments on them.

For example, Docker uses this OCI spec feature to restrict syscalls inside
containers. For the clone() syscall in particular, Docker restricts
the use of clone with CLONE_NEW* flags to clone into new namespaces inside
the container.

The system containers generated by the Sysbox runtime are more permissive in
terms on allowed syscalls, given that they always use the Linux
user-namespace, which ensures root privileges inside the container are
restricted to resources associated with the container.

This change causes Sysbox to remove syscall argument restrictions on
syscalls that Sysbox allows.

Prior to this change, cloning into new namespaces inside a sys container
launched by Docker was not possible. After this change, this is now possible.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>