Enable new grpc exchange to collect file-system state from sysbox-mgr.
* Abilitate new grpc exchange to collect file-system state from sysbox-mgr

Our goal here is to allow sysbox-runc to extract any file-system state that may
be required for the proper operation of a sys container. Even though this new
logic is generic enough (i.e. to collect any file, dir, softlink), the purpose
of this PR is to abilitate the creation of a softlink in the container's rootfs
to allow linux-kernel-headers to be properly found within any sys container
(prior to this, kernel-headers could be found in the expected path only in
Ubuntu sys-containers).

Example (host running Ubuntu Focal and Fedora sys-container):

* The following bind-mounts are created to address linux-headers exposure
  within the sys container (existing implementation):

[root@c3aa727de477 /]# findmnt
...
├─/usr/src/linux-headers-5.4.0-48                            /usr/src/linux-headers-5.4.0-48                    shiftfs  ro,relatime
├─/usr/src/linux-headers-5.4.0-48-generic                    /usr/src/linux-headers-5.4.0-48-generic            shiftfs  ro,relatime
└─/usr/lib/modules/5.4.0-48-generic                          /lib/modules/5.4.0-48-generic                      shiftfs  ro,relatime

* However, notice that this (/usr/src/linux-headers-<blah>") is not the path
  expected by Fedora distro (/usr/src/kernels/<blah>). That's why we now create
  a soft-link between the expected kernel-headers path and the existing one:

[root@c3aa727de477 /]# ls -lrt /usr/src/kernels/5.4.0-48-generic
lrwxrwxrwx 1 root root 39 Nov  5 05:20 /usr/src/kernels/5.4.0-48-generic -> /usr/src/linux-headers-5.4.0-48-generic
[root@c3aa727de477 /]#

Signed-off-by: Rodny Molina <rmolina@nestybox.com>