Disallow bind-mounts from nested host dirs in some cases.
When using uid shifting, if the mount is on a special dir of the sys
container (e.g., /var/lib/docker, /var/lib/kubelet, etc), and the
source of that mount is a child dir of another bind-mount source, we
must disallow it.

For example, mounting /a/b/c->/var/lib/docker is not allowed if /, /a,
or /a/b is also bind-mounted into the container. The reason is that
this requires mounting shiftfs on the parent path, which means that
the special mount would also be under shiftfs, and the latter is
something we want to avoid (as k8s for example will fail when shiftfs
is mounted on /var/lib/kubelet).

Signed-off-by: ctalledo <ctalledo@nestybox.com>