Fix ownership of container init process stdio in detached mode.
The container's stdio ownership must match the uid(gid) of the contaner's init
process.

Prior to this change, this was not the case when launching sys containers with
docker (without userns-remap). That's because containerd-shim was setting the
stdio ownership to root (since Docker thinks the container will have uid(gid)
set to root), but sysbox then allocates a non-root uid(gid) for the container.

Refer to sysbox issue #316 and #707 for full details.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>